fix query version, edit collections and cleanup
