+++ src/app/views/common/lazy-loader/component-registry/index.tsx
@@ -2,7 +2,7 @@ import LazyResponseHeadersV9 from '../../../query-response/headers/ResponseHeade
 import LazySnippetsV9 from '../../../query-response/snippets/SnippetsV9';
 import { default as LazyCopyButtonV9 } from '../../copy-button/CopyButtonV9';
 import LazySpecificPermissions from  '../../../query-runner/request/permissions';
-import LazyStatusMessages from  '../../../app-sections/StatusMessagesV9';
+import LazyStatusMessages from  '../../../app-sections/StatusMessages';
 import LazyResponseHeaders from  '../../../query-response/headers/ResponseHeaders';
 import LazyGraphToolkit from  '../../../query-response/graph-toolkit/GraphToolkit';
 import LazyCopyButton from  '../../copy-button/CopyButtonV9';
