+++ src/app/views/query-response/index.ts
@@ -1,4 +1,3 @@
 import QueryResponse from './QueryResponse';
-import { default as QueryResponseV9 } from './QueryResponseV9';
 
-export { QueryResponse, QueryResponseV9 };
+export { QueryResponse };
