+++ src/app/views/query-response/response/Response.tsx
@@ -1,39 +1,50 @@
+import { makeStyles, tokens } from '@fluentui/react-components';
+import { useAppSelector } from '../../../../store';
+import { CustomBody, ResponseBody } from '../../../../types/query-response';
+import { getContentType } from '../../../services/actions/query-action-creator-util';
+import ResponseDisplay from './ResponseDisplay';
+import { ResponseMessages } from './ResponseMessages';
 
+const useStyles = makeStyles({
+  container: {
+    display: 'flex',
+    flexDirection: 'column',
+    flex: 1,
+    height: '100%',
+    overflow: 'hidden'
+  },
+  messageBars: {
+    display: 'flex',
+    flexDirection: 'column',
+    gap: tokens.spacingHorizontalMNudge,
+    paddingTop: tokens.spacingHorizontalMNudge,
+    paddingBottom: tokens.spacingHorizontalMNudge
+  }
+});
 
-// import { useAppSelector } from '../../../../store';
-// import { getContentType } from '../../../services/actions/query-action-creator-util';
-// import {
-//   convertVhToPx, getResponseEditorHeight,
-//   getResponseHeight
-// } from '../../common/dimensions/dimensions-adjustment';
-// import ResponseDisplay from './ResponseDisplay';
-// import { ResponseMessages } from './ResponseMessages';
+const Response = () => {
+  const styles = useStyles();
+  const body = useAppSelector<ResponseBody>(
+    (state) => state.graphResponse.response.body
+  );
+  const headers = useAppSelector(
+    (state) => state.graphResponse.response.headers
+  );
 
-// const Response = () => {
-//   const response = useAppSelector((state) => state.dimensions.response);
-//   const body = useAppSelector((state) => state.graphResponse.response.body);
-//   const headers = useAppSelector((state) => state.graphResponse.response.headers);
-//   const responseAreaExpanded = useAppSelector((state) => state.responseAreaExpanded);
+  const contentDownloadUrl = (body as CustomBody)?.contentDownloadUrl;
+  const throwsCorsError = (body as CustomBody)?.throwsCorsError;
+  const contentType = getContentType(headers);
 
-//   const defaultHeight = convertVhToPx(getResponseHeight(response.height, responseAreaExpanded), 220);
-//   const monacoHeight = getResponseEditorHeight(150);
+  return (
+    <div className={styles.container}>
+      <div className={styles.messageBars}>
+        <ResponseMessages />
+      </div>
+      {!contentDownloadUrl && !throwsCorsError && headers && (
+        <ResponseDisplay contentType={contentType} body={body as string} />
+      )}
+    </div>
+  );
+};
 
-//   const contentDownloadUrl = body?.contentDownloadUrl;
-//   const throwsCorsError = body?.throwsCorsError;
-//   const contentType = getContentType(headers);
-
-//   return (
-//     <div style={{ display: 'block' }}>
-//       <ResponseMessages />
-//       {!contentDownloadUrl && !throwsCorsError && headers &&
-//         <ResponseDisplay
-//           contentType={contentType}
-//           body={body}
-//           height={responseAreaExpanded ? defaultHeight : monacoHeight}
-//         />}
-//     </div>
-//   );
-
-// };
-
-// export default Response;
+export default Response;
