+++ src/app/views/query-response/response/index.ts
@@ -1,4 +1,3 @@
-// import Response from './Response';
-import ResponseV9 from './ResponseV9';
-export { ResponseV9 };
+import Response from './Response';
+export { Response };
