--- src/app/views/query-runner/query-input/QueryInput.styles.ts
@@ -1,17 +1,0 @@
-import { ITheme } from '@fluentui/react';
-
-export const queryInputStyles = (theme: ITheme) => {
-  const controlWidth = '90%';
-  return {
-    autoComplete: {
-      input: {
-        minWidth: controlWidth,
-        overflowY: 'visible'
-      },
-      noSuggestions: {
-        color: theme.palette.black,
-        padding: 10
-      }
-    }
-  };
-};
