+++ src/app/views/query-runner/query-runner.scss
@@ -1,12 +1,5 @@
 @import "../../../styles/variables";
 
-.run-query-button {
-  button {
-    width: 100%;
-    padding: 3;
-  }
-}
-
  .ms-Dropdown:focus {
    outline-style:double !important;
    outline-width:thick !important ;
