+++ src/app/views/sidebar/history/History.tsx
@@ -209,6 +209,7 @@ const HistoryItems = (props: HistoryProps)=>{
   };
 
   const handleViewQuery = (query: IHistoryItem)=>{
+    console.log('query', query)
     const { sampleUrl, queryVersion } = parseSampleUrl(query.url);
     const sampleQuery: IQuery = {
       sampleUrl,
