+++ src/app/views/sidebar/sample-queries/SampleQueries.styles.ts
@@ -1,0 +1,32 @@
+import { makeStyles, tokens } from '@fluentui/react-components';
+export const useStyles = makeStyles({
+  container: {
+    marginTop: '6px'
+  },
+  searchBox: {
+    width: '100%',
+    maxWidth: '100%'
+  },
+  iconBefore: {
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: '2px'
+  },
+  badge: {
+    maxWidth: '50px'
+  },
+  disabled: {
+    backgroundColor: tokens.colorSubtleBackgroundHover,
+    '&:hover': {
+      cursor: 'not-allowed'
+    }
+  },
+  sampleQueries: {
+    height: 'calc(100vh - 374px)',
+    overflowY: 'auto'
+  },
+  itemLayout: {
+    paddingLeft: tokens.spacingHorizontalXXL
+  }
+});
