+++ src/types/submit-button.ts
@@ -2,7 +2,6 @@ export interface ISubmitButtonControl {
   handleOnClick: Function;
   submitting: boolean;
   text: string;
-  className: string;
   ariaLabel?: string;
   role?: string;
   disabled?: boolean;
